requirements.txt
# Python 3.12.3 compatible dependency ranges

# Core ML
numpy>=1.26,<3
pandas>=2.2,<3
scikit-learn>=1.4,<2
scipy>=1.12,<2
pyarrow>=15,<22

# Experiment tracking
mlflow>=2.11,<4

# Orchestration
apache-airflow>=2.10,<3
apache-airflow-providers-amazon>=8.11,<11
apache-airflow-providers-postgres>=5.12,<7

# App / serving
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1

# Config & utils
pyyaml>=6,<7
python-dotenv>=1.0,<2
pydantic>=2.6,<3
requests>=2.31,<3

# Dev/test
pytest>=8,<9
black>=24,<26
ruff>=0.3,<1
